centos7/dracut: ensure ext4 is loaded
OBS now always appends "ext4.unsupported=1" which leads to a fatal
modprobe error, because this is a SUSE-only parameter.
To work around this issue, manually load the ext4 module and its
dependencies with insmod, so that the root file system of the build VM
which is ext3, can be mounted.